adicionado validação, ordenação e estilização de login

+++ public/js/admin/excluirEvento.js
@@ -1,12 +1,19 @@
-function excluirEvento(evento_cad){
-    fetch(`/admin/excluirEvento/${evento_cad}`, {
-        method: "DELETE"
-    })
-    .then((r) => r.json())
-    .then((r) => {
-        alert(r.msg);
-        if(r.ok){
-            window.location.reload();
-        }
-    });
+
+function excluirEvento(evento_cad) {
+    // Exibe um diálogo de confirmação
+    let confirmacao = confirm("Tem certeza que deseja excluir este evento?");
+
+    // Se o usuário clicar em "OK" no diálogo de confirmação
+    if (confirmacao) {
+        fetch(`/admin/excluirEvento/${evento_cad}`, {
+            method: "DELETE"
+        })
+        .then((r) => r.json())
+        .then((r) => {
+            alert(r.msg);
+            if (r.ok) {
+                window.location.reload();
+            }
+        });
+    }
 }
