+++ public/js/admin/excluirPatrimonio.js
@@ -1,12 +1,16 @@
 function excluirPatrimonio(pat_etiqueta){
-    fetch(`/admin/excluirPatrimonio/${pat_etiqueta}`, {
-        method: "DELETE"
-    })
-    .then((r) => r.json())
-    .then((r) => {
-        alert(r.msg);
-        if(r.ok){
-            window.location.reload();
-        }
-    });
+    let confirmacao = confirm("Tem certeza que deseja excluir este evento?");
+    if(confirmacao){
+
+        fetch(`/admin/excluirPatrimonio/${pat_etiqueta}`, {
+            method: "DELETE"
+        })
+        .then((r) => r.json())
+        .then((r) => {
+            alert(r.msg);
+            if(r.ok){
+                window.location.reload();
+            }
+        });
+    }
 }
